requirements.txt
cryptography
Flask==3.0.2
flask-restx==1.3.0
Flask-httpauth==4.8.0
flask-sock==0.7.0
PyJWT==2.10.1
Flask-JWT-Extended==4.7.1 
numa
portainer-api
psutil
py-cpuinfo
pyyaml
requests
six
tinydb
virtfs
tcconfig
Werkzeug==3.0.1
pydantic==2.5.3
ansible-tower-cli==3.3.8
pytest-xprocess
kubernetes==29.0.0
openapi-slurm-0.0.38-client
sense-o-api==1.37
